refactor: split ContentView into its own file
Move the ContentView struct and EmptyStateView from MainView.swift into
a new ContentView.swift to reduce MainView.swift from 587 to ~356 lines
and give the main editing UI its own home. Pure source-level split with
no behavior changes.
